Keep attempting auto-trim until fully trimmed
Auto-trim is activated by editor changes. If auto-trim does not remove
some whitespace because the cursor is next to it, we want to remove it
after the cursor position changes, not wait until the next edit. To
achieve that, we keep the auto-trim active (firing at the debounce
interval) until fully trimmed.